+++ static/__tests__/dtale/reshape/DataViewer-reshape-transpose-test.jsx
@@ -92,7 +92,7 @@ describe("DataViewer tests", () => {
 
     setTimeout(() => {
       result.update();
-      clickMainMenuButton(result, "Reshape");
+      clickMainMenuButton(result, "Summarize Data");
       setTimeout(() => {
         result.update();
         result
@@ -167,7 +167,7 @@ describe("DataViewer tests", () => {
 
     setTimeout(() => {
       result.update();
-      clickMainMenuButton(result, "Reshape");
+      clickMainMenuButton(result, "Summarize Data");
       setTimeout(() => {
         result.update();
         t.equal(result.find(Reshape).length, 1, "should show reshape");