+++ tests/conftest.py
@@ -113,7 +113,18 @@ def custom_data(request):
 def state_data():
     df = pd.read_csv('{}/data/state-codes.csv'.format(os.path.dirname(__file__)))
     df.loc[:, 'val'] = df.index.values
+    df.loc[:, 'cat'] = np.random.uniform(1, 20, len(df))
     return df
+
+
+@pytest.fixture(scope="module")
+def scattergeo_data():
+    return pd.DataFrame({
+        'lat': np.random.uniform(-40, 40, 50),
+        'lon': np.random.uniform(-40, 40, 50),
+        'cat': np.random.uniform(1, 20, 50),
+        'val': np.random.randint(0, high=100, size=50)
+    })
 
 
 @pytest.fixture(scope="module")